pyproject.toml
[project]
name = "mol-gru4rec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "torch",
    "tensorflow-macos>=2.16.2",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "jupyterlab>=4.4.0",
    "isort>=6.0.1",
    "black>=25.1.0",
    "autoflake>=2.3.1",
    "flake8>=7.2.0",
    "google-cloud-storage>=2.19.0",
    "tqdm>=4.67.1",
    "apache-beam[gcp]>=2.64.0",
    "scikit-learn>=1.6.1",
]
